dagster_demo/requirements.txt
# requirements.txt
-r requirements-dev.txt          # dev / lint extras

dagster==1.11.2
dagster-webserver==1.11.2
dagster-mysql==0.27.2
dagster-msteams==0.27.2

click>=8,<8.2
sqlalchemy>=2.0
pymysql>=1.0
pandas>=2.2,<2.3
great_expectations==0.18.22
python-dotenv>=1.0
